requirements.txt
amqp==5.3.1 ; python_version >= "3.11" and python_version < "4.0"
annotated-types==0.7.0 ; python_version >= "3.11" and python_version < "4.0"
anyio==4.9.0 ; python_version >= "3.11" and python_version < "4.0"
async-timeout==5.0.1 ; python_version >= "3.11" and python_full_version < "3.11.3"
asyncpg==0.30.0 ; python_version >= "3.11" and python_version < "4.0"
bcrypt==4.3.0 ; python_version >= "3.11" and python_version < "4.0"
billiard==4.2.1 ; python_version >= "3.11" and python_version < "4.0"
celery==5.5.3 ; python_version >= "3.11" and python_version < "4.0"
cffi==1.17.1 ; python_version >= "3.11" and python_version < "4.0" and platform_python_implementation != "PyPy"
click-didyoumean==0.3.1 ; python_version >= "3.11" and python_version < "4.0"
click-plugins==1.1.1.2 ; python_version >= "3.11" and python_version < "4.0"
click-repl==0.3.0 ; python_version >= "3.11" and python_version < "4.0"
click==8.2.1 ; python_version >= "3.11" and python_version < "4.0"
colorama==0.4.6 ; python_version >= "3.11" and python_version < "4.0" and (platform_system == "Windows" or sys_platform == "win32")
cryptography==45.0.5 ; python_version >= "3.11" and python_version < "4.0"
dnspython==2.7.0 ; python_version >= "3.11" and python_version < "4.0"
ecdsa==0.19.1 ; python_version >= "3.11" and python_version < "4.0"
email-validator==2.2.0 ; python_version >= "3.11" and python_version < "4.0"
fastapi==0.115.13 ; python_version >= "3.11" and python_version < "4.0"
greenlet==3.2.3 ; python_version >= "3.11" and python_version < "3.14" and (platform_machine == "aarch64" or platform_machine == "ppc64le" or platform_machine == "x86_64" or platform_machine == "amd64" or platform_machine == "AMD64" or platform_machine == "win32" or platform_machine == "WIN32")
h11==0.16.0 ; python_version >= "3.11" and python_version < "4.0"
httptools==0.6.4 ; python_version >= "3.11" and python_version < "4.0"
idna==3.10 ; python_version >= "3.11" and python_version < "4.0"
kombu==5.5.4 ; python_version >= "3.11" and python_version < "4.0"
packaging==25.0 ; python_version >= "3.11" and python_version < "4.0"
passlib==1.7.4 ; python_version >= "3.11" and python_version < "4.0"
prompt-toolkit==3.0.51 ; python_version >= "3.11" and python_version < "4.0"
psycopg-binary==3.2.9 ; python_version >= "3.11" and python_version < "4.0" and implementation_name != "pypy"
psycopg==3.2.9 ; python_version >= "3.11" and python_version < "4.0"
pyasn1==0.6.1 ; python_version >= "3.11" and python_version < "4.0"
pycparser==2.22 ; python_version >= "3.11" and python_version < "4.0" and platform_python_implementation != "PyPy"
pydantic-core==2.33.2 ; python_version >= "3.11" and python_version < "4.0"
pydantic-settings==2.10.1 ; python_version >= "3.11" and python_version < "4.0"
pydantic==2.11.7 ; python_version >= "3.11" and python_version < "4.0"
python-dateutil==2.9.0.post0 ; python_version >= "3.11" and python_version < "4.0"
python-dotenv==1.1.1 ; python_version >= "3.11" and python_version < "4.0"
python-jose==3.5.0 ; python_version >= "3.11" and python_version < "4.0"
pyyaml==6.0.2 ; python_version >= "3.11" and python_version < "4.0"
redis==6.2.0 ; python_version >= "3.11" and python_version < "4.0"
rsa==4.9.1 ; python_version >= "3.11" and python_version < "4.0"
six==1.17.0 ; python_version >= "3.11" and python_version < "4.0"
sniffio==1.3.1 ; python_version >= "3.11" and python_version < "4.0"
sqlalchemy==2.0.41 ; python_version >= "3.11" and python_version < "4.0"
starlette==0.46.2 ; python_version >= "3.11" and python_version < "4.0"
typing-extensions==4.14.0 ; python_version >= "3.11" and python_version < "4.0"
typing-inspection==0.4.1 ; python_version >= "3.11" and python_version < "4.0"
tzdata==2025.2 ; python_version >= "3.11" and python_version < "4.0"
uvicorn==0.34.3 ; python_version >= "3.11" and python_version < "4.0"
uvloop==0.21.0 ; python_version >= "3.11" and python_version < "4.0" and sys_platform != "win32" and sys_platform != "cygwin" and platform_python_implementation != "PyPy"
vine==5.1.0 ; python_version >= "3.11" and python_version < "4.0"
watchfiles==1.1.0 ; python_version >= "3.11" and python_version < "4.0"
wcwidth==0.2.13 ; python_version >= "3.11" and python_version < "4.0"
websockets==15.0.1 ; python_version >= "3.11" and python_version < "4.0"
